Implementacion Manejo de errores con slf4j y correcciones de logger

+++ src/main/java/com/QuickOrder360/cliente/service/ClienteService.java
@@ -2,6 +2,7 @@ package com.QuickOrder360.cliente.service;
 
 import com.QuickOrder360.cliente.model.Cliente;
 import com.QuickOrder360.cliente.repository.ClienteRepository;
+import com.QuickOrder360.exception.ResourceNotFoundException;
 import jakarta.transaction.Transactional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,24 +20,27 @@ public class ClienteService {
     @Autowired
     private ClienteRepository clienteRepository;
 
-    public List<Cliente> findAll(){
+    public List<Cliente> findAll() {
         log.info("Buscando todos los clientes");
         return clienteRepository.findAll();
     }
 
-    public  Cliente save(Cliente cliente){
+    public Cliente save(Cliente cliente) {
         log.info("Guardando nuevo cliente");
         return clienteRepository.save(cliente);
     }
 
-    public void delete(Long id){
+    public void delete(Long id) {
         log.info("Eliminando cliente por ID: {}", id);
+        if (!clienteRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Cliente", id);
+        }
         clienteRepository.deleteById(id);
     }
 
-    public Cliente findById(Long id){
-        return clienteRepository.findById(id).get();
+    public Cliente findById(Long id) {
+        log.info("Buscando cliente por ID: {}", id);
+        return clienteRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Cliente", id));
     }
 }
-
-//l
